==> proj2/code/Step_02_Dataset_Creator.ipynb <==
"""
Cell For Papermill Parameters
"""

#Destinations Folders
pso_folder = "../../dataset/src_US/pos"
neg_folder = "../../dataset/src_US/neg"

#Target Flags
labels_file = "../data/TCIA Biopsy Data_2020-07-14.xlsx"

#Source Images Path
folder_path = "../../prostate_cancer_data/"
file_name = 'US_fnl.jpg'

import os
import pandas as pd
import glob
from pathlib import Path
from shutil import copyfile

def get_labels(patient_id):
    
    """
    get_labels return 1/0 for cancer or no cancer per patient
    
    :param patient_id: patientID
    :return: 0/1 for cancer or no cancer
    """ 
    
    df = labels_df[labels_df['Patient Number'].str.contains(patient_id)]
    
    df = df[df["% Cancer in Core"] > 0]
    
    if len(df) > 0:
        return 1
    else:
        return 0

def move_image(file_name,image,patient_id,scan_date):
    
    """
    move_image moves image from one folder to another
    
    :param file_name: file name
    :param image: image
    :param patient_id: patientid
    :param scan_date: scan date
    """     
        
    print("============================================")
    print(image)
    print(patient_id)
    print(scan_date)
    new_image_name = file_name +"_"+ scan_date +"_"+ patient_id +".jpg"
    print(new_image_name)
    
    if get_labels(patient_id) == 1:
        dst_folder = pso_folder
    else:
        dst_folder = neg_folder
    
    final_image = os.path.join(dst_folder, new_image_name)
    print(final_image)
            
    copyfile(image, final_image)

# ==========================================================================================
# Remove old Final Files
# ==========================================================================================
fileList = glob.glob(pso_folder+"/" + '*jpg*')

# Iterate over the list of filepaths & remove each file.
for filePath in fileList:
    try:
        os.remove(filePath)
    except:
        print("Error while deleting file : ", filePath)

fileList = glob.glob(neg_folder+"/" + '*jpg*')

# Iterate over the list of filepaths & remove each file.
for filePath in fileList:
    try:
        os.remove(filePath)
    except:
        print("Error while deleting file : ", filePath)

labels_df = pd.read_excel (labels_file)
fileList = glob.glob(folder_path + '**/*' +file_name  , recursive=True)

print(len(fileList))

for image in fileList:
    patient_id = image[119:123]
    scan_date = image[124:133]
    print(patient_id)
    print(scan_date)
    move_image(file_name,image,patient_id,scan_date)
